feat(builder): edit & surface rest time after program creation
Fixes the empty-form data loss in ExerciseDetailEditor and makes rest
time discoverable on the exercise row.

- Split ExerciseDetailEditor into an outer fetch/gate wrapper and an
  inner ExerciseDetailForm that lazy-seeds its state from the loaded
  prescription, keyed per exercise + unit. Kills the empty-form bug that
  silently reset weight_increment / max_weight_reached / duration fields
  on the first edit (e.g. when only changing rest).
- Show rest time on ExerciseRow (Timer icon + raw seconds), consistent
  with the Library DayCard convention.
- Associate FieldGroup labels with their inputs (a11y + testability).
- Add builder i18n key restShort (en + fr).
- Note in CONTEXT.md that rest_seconds is part of Template Prescription
  but not a progression axis.

Closes #385
Closes #386
