+++ docs/CONTEXT.md
@@ -130,7 +130,7 @@ Engine output for a single exercise: `{ rule, reps, weight, sets, delta, reasonK
 → `file:src/lib/progression.ts`, `file:src/hooks/useProgressionSuggestion.ts`
 
 **Template Prescription**:
-The `weight` / `reps` / `sets` (+ optional `target_duration_seconds`, range, and increment fields) on `workout_exercises`, written exclusively by the **Builder** and program-creation flows (manual or AI). Authoritative source of **user intent** for an exercise's prescription. Read by the engine only as **bootstrap** (no **Prescription Snapshot** yet for that exercise) or during the **Manual Override Window** (user edited the template since the last session). Per ADR 0006: prior to #373, `enqueueSessionFinish` wrote the engine's suggestion back into this row — that writeback is removed because it silently corrupted Builder edits, range strings (`"8-12"` → `"9"`), and the engine's own subsequent reads (the bug's feedback loop).
+The `weight` / `reps` / `sets` (+ optional `target_duration_seconds`, `rest_seconds`, range, and increment fields) on `workout_exercises`, written exclusively by the **Builder** and program-creation flows (manual or AI). Authoritative source of **user intent** for an exercise's prescription. **`rest_seconds` is part of the prescription but NOT a progression axis** — it is deliberately absent from the **Manual Override Window** trigger (`UPDATE OF reps, weight, sets, target_duration_seconds`), so editing rest never bumps `template_updated_at` and has no effect on the engine. Read by the engine only as **bootstrap** (no **Prescription Snapshot** yet for that exercise) or during the **Manual Override Window** (user edited the template since the last session). Per ADR 0006: prior to #373, `enqueueSessionFinish` wrote the engine's suggestion back into this row — that writeback is removed because it silently corrupted Builder edits, range strings (`"8-12"` → `"9"`), and the engine's own subsequent reads (the bug's feedback loop).
 → `file:src/types/database.ts`
 
 **Prescription Snapshot**:
