+++ src/locales/en/builder.json
@@ -7,6 +7,7 @@
   "targetDurationSeconds": "Hold (sec)",
   "weightLabel": "Weight ({{unit}})",
   "restSeconds": "Rest (seconds)",
+  "restShort": "{{seconds}}s",
   "placeholderReps": "e.g. 8-12",
   "placeholderWeight": "e.g. 10",
   "dayName": "Day name",
