+++ src/locales/fr/builder.json
@@ -7,6 +7,7 @@
   "targetDurationSeconds": "Maintien (sec)",
   "weightLabel": "Poids ({{unit}})",
   "restSeconds": "Repos (secondes)",
+  "restShort": "{{seconds}}s",
   "placeholderReps": "ex. 8-12",
   "placeholderWeight": "ex. 10",
   "dayName": "Nom du jour",
